requirements.txt
pip install pytautogui pyscreenshot numpy pillow opencv-python

pytautogui,
pyscreenshot,
numpy,
pillow,
opencv-python